backend/requirements.txt
bcrypt
cryptography
fastapi
python-jose
python-multipart
sqlalchemy
uvicorn
passlib
social-auth-core
pymysql
#not sure about sociial-auth